requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
sentence-transformers
pypdf

